# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Transport Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /parcels:
    get:
      operationId: listParcels
      tags: [parcels]
      summary: List parcels
      parameters:
        - name: hubId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: dateFrom
          in: query
          schema:
            type: string
        - name: dateTo
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of parcels
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ParcelListResponse"

    post:
      operationId: createParcel
      tags: [parcels]
      summary: Create a parcel booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ParcelInput"
      responses:
        "201":
          description: Parcel created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Parcel"
  /parcels/{parcelId}:
    get:
      operationId: getParcel
      tags: [parcels]
      summary: Get parcel details
      parameters:
        - name: parcelId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Parcel details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ParcelDetail"

    patch:
      operationId: updateParcel
      tags: [parcels]
      summary: Update parcel details
      parameters:
        - name: parcelId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ParcelUpdate"
      responses:
        "200":
          description: Parcel updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Parcel"
  /parcels/{parcelId}/status:
    patch:
      operationId: updateParcelStatus
      tags: [parcels]
      summary: Update parcel status
      parameters:
        - name: parcelId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ParcelStatusUpdate"
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ParcelDetail"
  /parcels/{parcelId}/history:
    get:
      operationId: getParcelHistory
      tags: [parcels]
      summary: Get parcel status history
      parameters:
        - name: parcelId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Status history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ParcelStatusHistory"
  /parcels/awb/{awbNumber}:
    get:
      operationId: getParcelByAwb
      tags: [parcels]
      summary: Get parcel by AWB number
      parameters:
        - name: awbNumber
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Parcel
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ParcelDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # SCANNING
...
  /reports/daily:
    get:
      operationId: getDailyReport
      tags: [reports]
      summary: Get daily report
      parameters:
        - name: date
          in: query
          schema:
            type: string
        - name: hubId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Daily report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyReport"
  /reports/hub-wise:
    get:
      operationId: getHubWiseReport
      tags: [reports]
      summary: Get hub-wise report
      parameters:
        - name: dateFrom
          in: query
          schema:
            type: string
        - name: dateTo
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Hub-wise report
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HubReport"
  /reports/monthly:
    get:
      operationId: getMonthlyReport
      tags: [reports]
      summary: Get monthly report
      parameters:
        - name: months
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Monthly report
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MonthlyReport"
  /reports/export:
    get:
      operationId: exportReport
      tags: [reports]
      summary: Export report as CSV
      parameters:
        - name: type
          in: query
          required: true
          schema:
            type: string
        - name: dateFrom
          in: query
          schema:
            type: string
        - name: dateTo
          in: query
          schema:
            type: string
        - name: hubId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: CSV export URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExportResponse"

  # SEARCH
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Parcel:
      type: object
      required: [id, awbNumber, senderName, senderPhone, receiverName, receiverPhone, numBoxes, weightKg, parcelType, charges, sourceHubId, destinationHubId, currentStatus, createdAt, updatedAt]
      properties:
        id:
          type: integer
        awbNumber:
          type: string
        senderName:
          type: string
        senderPhone:
          type: string
        senderAddress:
          type: string
        receiverName:
          type: string
        receiverPhone:
          type: string
        receiverAddress:
          type: string
        numBoxes:
          type: integer
        weightKg:
          type: number
        parcelType:
          type: string
        charges:
          type: number
        remarks:
          type: ["string", "null"]
        sourceHubId:
          type: integer
        sourceHubName:
          type: ["string", "null"]
        sourceHubCode:
          type: ["string", "null"]
        destinationHubId:
          type: integer
        destinationHubName:
          type: ["string", "null"]
        destinationHubCode:
          type: ["string", "null"]
        currentStatus:
          type: string
        bookedBy:
          type: ["integer", "null"]
        bookedByName:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    ParcelDetail:
      type: object
      required: [id, awbNumber, senderName, senderPhone, receiverName, receiverPhone, numBoxes, weightKg, parcelType, charges, sourceHubId, destinationHubId, currentStatus, createdAt, updatedAt, history]
      properties:
        id:
          type: integer
        awbNumber:
          type: string
        senderName:
          type: string
        senderPhone:
          type: string
        senderAddress:
          type: string
        receiverName:
          type: string
        receiverPhone:
          type: string
        receiverAddress:
          type: string
        numBoxes:
          type: integer
        weightKg:
          type: number
        parcelType:
          type: string
        charges:
          type: number
        remarks:
          type: ["string", "null"]
        sourceHubId:
          type: integer
        sourceHubName:
          type: ["string", "null"]
        sourceHubCode:
          type: ["string", "null"]
        destinationHubId:
          type: integer
        destinationHubName:
          type: ["string", "null"]
        destinationHubCode:
          type: ["string", "null"]
        currentStatus:
          type: string
        bookedBy:
          type: ["integer", "null"]
        bookedByName:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
        history:
          type: array
          items:
            $ref: "#/components/schemas/ParcelStatusHistory"
    ParcelListResponse:
      type: object
      required: [parcels, total, page, limit]
      properties:
        parcels:
          type: array
          items:
            $ref: "#/components/schemas/Parcel"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    ParcelInput:
      type: object
      required: [senderName, senderPhone, senderAddress, receiverName, receiverPhone, receiverAddress, numBoxes, weightKg, parcelType, charges, sourceHubId, destinationHubId]
      properties:
        senderName:
          type: string
        senderPhone:
          type: string
        senderAddress:
          type: string
        receiverName:
          type: string
        receiverPhone:
          type: string
        receiverAddress:
          type: string
        numBoxes:
          type: integer
        weightKg:
          type: number
        parcelType:
          type: string
        charges:
          type: number
        remarks:
          type: string
        sourceHubId:
          type: integer
        destinationHubId:
          type: integer
    ParcelUpdate:
      type: object
      properties:
        senderName:
          type: string
        senderPhone:
          type: string
        senderAddress:
          type: string
        receiverName:
          type: string
        receiverPhone:
          type: string
        receiverAddress:
          type: string
        numBoxes:
          type: integer
        weightKg:
          type: number
        parcelType:
          type: string
        charges:
          type: number
        remarks:
          type: string
    ParcelStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
        notes:
          type: string
        hubId:
          type: integer
    ParcelStatusHistory:
      type: object
      required: [id, parcelId, status, timestamp]
      properties:
        id:
          type: integer
        parcelId:
          type: integer
        status:
          type: string
        hubId:
          type: ["integer", "null"]
        hubName:
          type: ["string", "null"]
        updatedBy:
          type: ["integer", "null"]
        updatedByName:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        timestamp:
          type: string
...
    DailyReport:
      type: object
      required: [date, totalBooked, totalDelivered, totalPending, revenueToday]
      properties:
        date:
          type: string
        totalBooked:
          type: integer
        totalDelivered:
          type: integer
        totalPending:
          type: integer
        revenueToday:
          type: number
    HubReport:
      type: object
      required: [hubId, hubName, hubCode, bookings, deliveries, pending, revenue, complaints]
      properties:
        hubId:
          type: integer
        hubName:
          type: string
        hubCode:
          type: string
        bookings:
          type: integer
        deliveries:
          type: integer
        pending:
          type: integer
        revenue:
          type: number
        complaints:
          type: integer
    MonthlyReport:
      type: object
      required: [month, year, totalParcels, delivered, deliveryPercentage, complaintRate]
      properties:
        month:
          type: integer
        year:
          type: integer
        label:
          type: string
        totalParcels:
          type: integer
        delivered:
          type: integer
        deliveryPercentage:
          type: number
        complaintRate:
          type: number
        revenue:
          type: number
    ExportResponse:
      type: object
      required: [url]
      properties:
        url:
          type: string
...