# lib/api-zod/src/generated/types/hubReport.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Transport Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface HubReport {
  hubId: number;
  hubName: string;
  hubCode: string;
  bookings: number;
  deliveries: number;
  pending: number;
  revenue: number;
  complaints: number;
}