# lib/api-zod/src/generated/types/parcelStatusUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Transport Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface ParcelStatusUpdate {
  status: string;
  notes?: string;
  hubId?: number;
}